+++ app/species/cache.py
@@ -1,10 +1,11 @@
-from functools import lru_cache
+from cachetools import cached, LRUCache
+from cachetools.keys import hashkey
 
 from fastapi import APIRouter, HTTPException, status
 from sqlmodel import Session, func, select, delete
 
 from app.database import DbDependency
-from app.settings_env import EnvSettings
+from app.settings_env import EnvDependency, EnvSettings
 from app.sqlmodels import Taxon
 import app.species.indicia as driver
 from app.utility.search import Search
@@ -80,12 +81,38 @@ async def delete_cache_item(
     response_model=Taxon)
 async def read_taxon_by_tvk(
         db: DbDependency,
+        env: EnvDependency,
         tvk: str):
-    return get_taxon_by_tvk(db, tvk)
+    # This endpoint is different from /species/taxon_by_tvk as it returns
+    # additional information, including the cache id.
+    try:
+        return get_taxon_by_tvk(db, env, tvk)
+    except ValueError as e:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=str(e))
+    except Exception as e:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail=str(e))
 
 
-@lru_cache(maxsize=1024)
 def get_taxon_by_tvk(db: Session, env: EnvSettings, tvk: str) -> Taxon:
+    """Look up the taxon with given TVK.
+
+    Wrapper for _get_taxon_by_tvk_wrapped so that exceptions can be cached.
+    """
+
+    taxon = _get_taxon_by_tvk_wrapped(db, env, tvk)
+    if isinstance(taxon, Exception):
+        raise taxon
+    else:
+        # Return the deserialized taxon
+        return Taxon(**taxon)
+
+
+@cached(cache=LRUCache(maxsize=1024),  key=lambda db, env, tvk: hashkey(tvk))
+def _get_taxon_by_tvk_wrapped(db: Session, env: EnvSettings, tvk: str) -> dict:
     """Look up the taxon with given TVK."""
 
     # First check our local database
@@ -95,12 +122,24 @@ def get_taxon_by_tvk(db: Session, env: EnvSettings, tvk: str) -> Taxon:
     ).first()
     if not taxon:
         # If not found, add from the remote database.
-        taxon = add_taxon_by_tvk(db, env, tvk)
-    return taxon
+        taxon = _add_taxon_by_tvk(db, env, tvk)
+
+    if isinstance(taxon, Exception):
+        return taxon
+    else:
+        # Return the serialized taxon.
+        # This is a workaround to a problem returning the taxon as a model.
+        # There seemed to be a link between the model and the database
+        # session which meant the cache response was invalid in a different
+        # session if I cached the model. Not well understood.
+        return taxon.model_dump()
 
 
-def add_taxon_by_tvk(db: Session,  env: EnvSettings, tvk: str) -> Taxon:
-    """Look up taxon with given TVK and add to cache."""
+def _add_taxon_by_tvk(db: Session,  env: EnvSettings, tvk: str) -> Taxon:
+    """Look up taxon with given TVK and add to cache.
+
+    Exceptions are returned so that they can be cached.
+     """
     params = {
         'search_code': tvk,
         'include': '["data"]'
@@ -109,7 +148,7 @@ def add_taxon_by_tvk(db: Session,  env: EnvSettings, tvk: str) -> Taxon:
     taxa = driver.parse_response_taxa(response)
 
     if len(taxa) == 0:
-        raise ValueError(f"TVK {tvk} not recognised.")
+        return ValueError(f"TVK {tvk} not recognised.")
     else:
         taxon = taxa[0]
         db.add(taxon)
@@ -118,9 +157,23 @@ def add_taxon_by_tvk(db: Session,  env: EnvSettings, tvk: str) -> Taxon:
         return taxon
 
 
-@lru_cache(maxsize=1024)
 def get_taxon_by_name(db: Session, env: EnvSettings, name: str) -> Taxon:
-    """Look up taxon with given name."""
+    """Look up taxon with given name.
+
+    Wrapper for _get_taxon_by_name_wrapped so that exceptions can be cached.
+    """
+
+    taxon = _get_taxon_by_name_wrapped(db, env, name)
+    if isinstance(taxon, Exception):
+        raise taxon
+    else:
+        # Return the deserialized taxon
+        return Taxon(**taxon)
+
+
+@cached(cache=LRUCache(maxsize=1024), key=lambda db, env, name: hashkey(name))
+def _get_taxon_by_name_wrapped(db: Session, env: EnvSettings, name: str) -> dict:
+    """Look up taxon with given name in local database."""
 
     search_name = Search.get_search_name(name)
     # First check our local database.
@@ -129,23 +182,49 @@ def get_taxon_by_name(db: Session, env: EnvSettings, name: str) -> Taxon:
     ).first()
     if not taxon:
         # If not found, add from the remote database.
-        taxon = add_taxon_by_name(db, env, name)
-    return taxon
+        taxon = _add_taxon_by_name(db, env, name)
+
+    if isinstance(taxon, Exception):
+        return taxon
+    else:
+        # Return the serialized taxon.
+        # This is a workaround to a problem returning the taxon as a model.
+        # There seemed to be a link between the model and the database
+        # session which meant the cache response was invalid in a different
+        # session if I cached the model. Not well understood.
+        return taxon.model_dump()
 
 
-def add_taxon_by_name(db: Session, env: EnvSettings, name: str) -> Taxon:
-    """Look up taxon and add to cache."""
+def _add_taxon_by_name(db: Session, env: EnvSettings, name: str) -> Taxon:
+    """Look up taxon in remote database and add to local.
+
+    Exceptions are returned so that they can be cached.
+    """
     params = {
         'searchQuery': name,
-        'include': '["data"]'
+        'include': '["data"]',
+        'wholeWords': 'true',
     }
     response = driver.make_search_request(env, params)
     taxa = driver.parse_response_taxa(response)
 
     if len(taxa) == 0:
-        raise ValueError("Name not recognised.")
+        return ValueError(f"Name '{name}' not recognised.")
     else:
+        # Use the first suggestion. This may be naive but Indicia does seem to
+        # sort the results in a helpful way.
         taxon = taxa[0]
+        # The name returned is not always the same as the one supplied.
+        # E.g. A search for 'Tern' returns 'Tern species'.
+        request_search_name = Search.get_search_name(name)
+        response_search_name = Search.get_search_name(taxon.name)
+        if request_search_name != response_search_name:
+            error = f"Name '{name}' not recognised. "
+            suggestions = ', '.join(taxon.name for taxon in taxa)
+            error += f"Suggestions: {suggestions}"
+            return ValueError(error)
+
+        # Add the new taxon to the cache.
         db.add(taxon)
         db.commit()
         db.refresh(taxon)