+++ app/verify/verify_routes.py
@@ -1,13 +1,14 @@
 import time
 
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter
 
-from app.auth import get_current_user
+from app.auth import UserDependency
 from app.database import DbDependency
 from app.rule.org_group.org_group_repo import OrgGroupRepo
 from app.rule.rule_repo import RuleRepo
-from app.settings_env import EnvDependency
+from app.settings import SettingsDependency
 import app.species.cache as cache
+from app.usage.usage_repo import UsageRepo
 from app.utility.sref.sref_factory import SrefFactory
 from app.utility.vague_date import VagueDate
 
@@ -16,7 +17,6 @@ from .verify_models import VerifyPack, Verified, VerifiedPack
 
 router = APIRouter(
     tags=["Verify"],
-    dependencies=[Depends(get_current_user)]
 )
 
 
@@ -27,7 +27,8 @@ router = APIRouter(
     response_model_exclude_none=True)
 async def verify(
     db: DbDependency,
-    env: EnvDependency,
+    settings: SettingsDependency,
+    user: UserDependency,
     data: VerifyPack,
     verbose: int = 1,
 ):
@@ -56,6 +57,7 @@ async def verify(
     # might offer more control. For now it is used as a boolean internally.
 
     start = time.time_ns()
+    env = settings.env
     results = []
     for record in data.records:
         # Our response begins with the input data.
@@ -136,8 +138,13 @@ async def verify(
 
     duration = time.time_ns() - start
 
+    repo = UsageRepo(db)
+    repo.update_verification_usage(user.name, len(results))
+
     return VerifiedPack(
         org_group_rules_list=data.org_group_rules_list,
         records=results,
-        duration_ns=duration
+        duration_ns=duration,
+        rules_commit=settings.db.rules_commit,
+        rules_update_time=settings.db.rules_update_time
     )