+++ test/rule/difficulty/test_difficulty.py
@@ -1,7 +1,11 @@
+import json
+
 from fastapi.testclient import TestClient
 from sqlmodel import Session
 
 from app.sqlmodels import OrgGroup, Taxon, DifficultyCode, DifficultyRule
+
+from ...mocks import mock_make_search_request
 
 
 class TestDifficulty:
@@ -9,7 +13,13 @@ class TestDifficulty:
 
     Fixtures for database and authentication come from ../conftest.py"""
 
-    def test_difficulty(self, client: TestClient):
+    def test_difficulty(self, client: TestClient, mocker):
+        # Mock the Indicia warehouse.
+        mocker.patch(
+            'app.species.indicia.make_search_request',
+            mock_make_search_request
+        )
+
         # Get database connection from client.
         engine = client.app.context['engine']
         with Session(engine) as db:
@@ -78,6 +88,12 @@ class TestDifficulty:
             assert result[0]['difficulty'] == 1
             assert result[0]['text'] == 'Easy'
 
+            # Request difficulty rules for invlaid tvk.
+            response = client.get('/rules/difficulty/tvk/ABC123')
+            assert response.status_code == 404
+            result = json.loads(response.text)
+            assert result['detail'] == 'TVK ABC123 not recognised.'
+
             # Request difficulty rules for organism_key.
             response = client.get(
                 f'/rules/difficulty/organism_key/{taxon.organism_key}')
@@ -87,3 +103,9 @@ class TestDifficulty:
             assert result[0]['group'] == 'group1'
             assert result[0]['difficulty'] == 1
             assert result[0]['text'] == 'Easy'
+
+            # Request difficulty rules for invlaid organism_key.
+            response = client.get('/rules/difficulty/organism_key/ABC123')
+            assert response.status_code == 200
+            result = response.json()
+            assert len(result) == 0