+++ test/verify/test_verify.py
@@ -1,7 +1,15 @@
+from datetime import datetime
+
 from fastapi.testclient import TestClient
 from sqlmodel import Session
 
-from app.sqlmodels import OrgGroup, Taxon, TenkmRule, DifficultyCode, DifficultyRule
+from app.sqlmodels import (
+    OrgGroup,
+    Taxon,
+    TenkmRule,
+    DifficultyCode,
+    DifficultyRule
+)
 from app.verify.verify_models import VerifyPack, OrgGroupRules
 
 from ..mocks import mock_make_search_request
@@ -13,6 +21,11 @@ class TestVerify:
         pack = VerifyPack(
             records=[],
         )
+        # Use settings from client context to set rule metadata.
+        settings = client.app.context['settings']
+        settings.db.rules_commit = 'test123'
+        settings.db.rules_update_time = '2026-02-23 16:59:59'
+
         response = client.post(
             '/verify',
             json=pack.model_dump(),
@@ -20,8 +33,14 @@ class TestVerify:
         assert response.status_code == 200
         verified = response.json()
         assert verified['records'] == []
+        assert verified['rules_commit'] == 'test123'
+        assert verified['rules_update_time'] == '2026-02-23 16:59:59'
 
     def test_valid_record(self, client: TestClient, mocker):
+        # Use settings from client context to set rule metadata.
+        settings = client.app.context['settings']
+        settings.db.rules_commit = 'test123'
+        settings.db.rules_update_time = '2026-02-23 16:59:59'
         # Get database connection from client.
         engine = client.app.context['engine']
         with Session(engine) as db:
@@ -47,15 +66,18 @@ class TestVerify:
                 json=pack.model_dump(),
             )
             assert response.status_code == 200
-            verified = response.json()['records'][0]
-            assert verified['name'] == "Adalia bipunctata"
-            assert verified['date'] == "03/04/2024"
-            assert verified['sref']['gridref'] == "TL123456"
-            assert verified['result'] == 'warn'
-            assert 'id_difficulty' not in verified
-            assert len(verified['messages']) == 1
-            assert verified['messages'][0] == (
+            verified = response.json()
+            record = verified['records'][0]
+            assert record['name'] == "Adalia bipunctata"
+            assert record['date'] == "03/04/2024"
+            assert record['sref']['gridref'] == "TL123456"
+            assert record['result'] == 'warn'
+            assert 'id_difficulty' not in record
+            assert len(record['messages']) == 1
+            assert record['messages'][0] == (
                 "No rules exist for this taxon.")
+            assert verified['rules_commit'] == 'test123'
+            assert verified['rules_update_time'] == '2026-02-23 16:59:59'
 
             # Update to test against specific org_group.
             rules = OrgGroupRules(
@@ -68,12 +90,15 @@ class TestVerify:
                 json=pack.model_dump(),
             )
             assert response.status_code == 200
-            verified = response.json()['records'][0]
-            assert verified['result'] == 'fail'
-            assert 'id_difficulty' not in verified
-            assert len(verified['messages']) == 1
-            assert verified['messages'][0] == (
+            verified = response.json()
+            record = verified['records'][0]
+            assert record['result'] == 'fail'
+            assert 'id_difficulty' not in record
+            assert len(record['messages']) == 1
+            assert record['messages'][0] == (
                 "Unrecognised organisation:group, 'UK Ladybird Survey:UKLS'.")
+            assert verified['rules_commit'] == 'test123'
+            assert verified['rules_update_time'] == '2026-02-23 16:59:59'
 
             # Create the missing org_group.
             org_group = OrgGroup(
@@ -87,12 +112,15 @@ class TestVerify:
                 json=pack.model_dump(),
             )
             assert response.status_code == 200
-            verified = response.json()['records'][0]
-            assert verified['result'] == 'warn'
-            assert 'id_difficulty' not in verified
-            assert len(verified['messages']) == 1
-            assert verified['messages'][0] == (
+            verified = response.json()
+            record = verified['records'][0]
+            assert record['result'] == 'warn'
+            assert 'id_difficulty' not in record
+            assert len(record['messages']) == 1
+            assert record['messages'][0] == (
                 "UK Ladybird Survey:UKLS: No rules exist for this taxon.")
+            assert verified['rules_commit'] == 'test123'
+            assert verified['rules_update_time'] == '2026-02-23 16:59:59'
 
             # Create the missing difficulty rule.
             # Create taxa.
@@ -135,14 +163,17 @@ class TestVerify:
                 json=pack.model_dump(),
             )
             assert response.status_code == 200
-            verified = response.json()['records'][0]
-            assert verified['result'] == 'warn'
-            assert verified['id_difficulty'] == 1
-            assert len(verified['messages']) == 2
-            assert verified['messages'][0] == (
+            verified = response.json()
+            record = verified['records'][0]
+            assert record['result'] == 'warn'
+            assert record['id_difficulty'] == 1
+            assert len(record['messages']) == 2
+            assert record['messages'][0] == (
                 "No rules run.")
-            assert verified['messages'][1] == (
+            assert record['messages'][1] == (
                 "UK Ladybird Survey:UKLS:difficulty:1: Easy")
+            assert verified['rules_commit'] == 'test123'
+            assert verified['rules_update_time'] == '2026-02-23 16:59:59'
 
             # Update to test against specific rule.
             pack.org_group_rules_list[0].rules = ['tenkm']
@@ -152,14 +183,17 @@ class TestVerify:
                 json=pack.model_dump(),
             )
             assert response.status_code == 200
-            verified = response.json()['records'][0]
-            assert verified['result'] == 'warn'
-            assert verified['id_difficulty'] == 1
-            assert len(verified['messages']) == 2
-            assert verified['messages'][0] == (
+            verified = response.json()
+            record = verified['records'][0]
+            assert record['result'] == 'warn'
+            assert record['id_difficulty'] == 1
+            assert len(record['messages']) == 2
+            assert record['messages'][0] == (
                 "No rules run.")
-            assert verified['messages'][1] == (
+            assert record['messages'][1] == (
                 "UK Ladybird Survey:UKLS:difficulty:1: Easy")
+            assert verified['rules_commit'] == 'test123'
+            assert verified['rules_update_time'] == '2026-02-23 16:59:59'
 
             # Create a tenkm to test against.
             # Create tenkm rule for org_group and taxon.
@@ -180,14 +214,17 @@ class TestVerify:
                 json=pack.model_dump(),
             )
             assert response.status_code == 200
-            verified = response.json()['records'][0]
-            assert verified['result'] == 'pass'
-            assert verified['id_difficulty'] == 1
-            assert len(verified['messages']) == 2
-            assert verified['messages'][0] == (
+            verified = response.json()
+            record = verified['records'][0]
+            assert record['result'] == 'pass'
+            assert record['id_difficulty'] == 1
+            assert len(record['messages']) == 2
+            assert record['messages'][0] == (
                 'Rules run: tenkm')
-            assert verified['messages'][1] == (
+            assert record['messages'][1] == (
                 'UK Ladybird Survey:UKLS:difficulty:1: Easy')
+            assert verified['rules_commit'] == 'test123'
+            assert verified['rules_update_time'] == '2026-02-23 16:59:59'
 
             # Now try again with verbose = 0.
             response = client.post(
@@ -196,12 +233,15 @@ class TestVerify:
                 json=pack.model_dump(),
             )
             assert response.status_code == 200
-            verified = response.json()['records'][0]
-            assert verified['result'] == 'pass'
-            assert verified['id_difficulty'] == 1
-            assert len(verified['messages']) == 1
-            assert verified['messages'][0] == (
+            verified = response.json()
+            record = verified['records'][0]
+            assert record['result'] == 'pass'
+            assert record['id_difficulty'] == 1
+            assert len(record['messages']) == 1
+            assert record['messages'][0] == (
                 'Rules run: tenkm')
+            assert verified['rules_commit'] == 'test123'
+            assert verified['rules_update_time'] == '2026-02-23 16:59:59'
 
             # Change record location to outside the tenkm rule - should fail.
             pack.records[0].sref.gridref = "TL 654 321"
@@ -210,15 +250,18 @@ class TestVerify:
                 json=pack.model_dump(),
             )
             assert response.status_code == 200
-            verified = response.json()['records'][0]
-            assert verified['result'] == 'fail'
-            assert verified['id_difficulty'] == 1
-            assert len(verified['messages']) == 2
-            assert verified['messages'][0] == (
+            verified = response.json()
+            record = verified['records'][0]
+            assert record['result'] == 'fail'
+            assert record['id_difficulty'] == 1
+            assert len(record['messages']) == 2
+            assert record['messages'][0] == (
                 'UK Ladybird Survey:UKLS:difficulty:1: Easy')
-            assert verified['messages'][1] == (
+            assert record['messages'][1] == (
                 "UK Ladybird Survey:UKLS:tenkm: Location is outside known "
                 "distribution.")
+            assert verified['rules_commit'] == 'test123'
+            assert verified['rules_update_time'] == '2026-02-23 16:59:59'
 
             # Remove the rule list - should still fail.
             pack.org_group_rules_list = []
@@ -227,12 +270,39 @@ class TestVerify:
                 json=pack.model_dump(),
             )
             assert response.status_code == 200
-            verified = response.json()['records'][0]
-            assert verified['result'] == 'fail'
-            assert verified['id_difficulty'] == 1
-            assert len(verified['messages']) == 2
-            assert verified['messages'][0] == (
+            verified = response.json()
+            record = verified['records'][0]
+            assert record['result'] == 'fail'
+            assert record['id_difficulty'] == 1
+            assert len(record['messages']) == 2
+            assert record['messages'][0] == (
                 'UK Ladybird Survey:UKLS:difficulty:1: Easy')
-            assert verified['messages'][1] == (
+            assert record['messages'][1] == (
                 "UK Ladybird Survey:UKLS:tenkm: Location is outside known "
                 "distribution.")
+            assert verified['rules_commit'] == 'test123'
+            assert verified['rules_update_time'] == '2026-02-23 16:59:59'
+
+            # Put two records in the pack.
+            record = pack.records[0]
+            pack.records.append(record)
+            response = client.post(
+                '/verify',
+                json=pack.model_dump(),
+            )
+            assert response.status_code == 200
+            verified = response.json()
+            assert len(verified['records']) == 2
+
+            # Check usage
+            year = datetime.now().year
+            response = client.get(f"/usage/{year}")
+            assert response.status_code == 200
+            usages = response.json()
+            assert len(usages) == 1
+            usage = usages[0]
+            assert usage['user_name'] == 'Tom'
+            assert usage['verification_requests'] == 10
+            assert usage['validation_requests'] == 0
+            assert usage['verification_records'] == 11
+            assert usage['validation_records'] == 0